+++ core/modules/product/index.ts
@@ -1,6 +1,7 @@
 import { productModule } from './store/product'
 import { attributeModule } from './store/attribute'
 import { stockModule } from './store/stock'
+import { taxModule } from './store/tax'
 import { VueStorefrontModule, VueStorefrontModuleConfig } from '@vue-storefront/module'
 
 export const KEY = 'product'
@@ -10,7 +11,8 @@ const moduleConfig: VueStorefrontModuleConfig = {
   store: { modules: [
     { key: KEY, module: productModule },
     { key: 'attribute', module: attributeModule },
-    { key: 'stock', module: stockModule }
+    { key: 'stock', module: stockModule },
+    { key: 'tax', module: taxModule }
   ] },
 }
