+++ core/modules/product/store/tax/mutation-types.ts
@@ -1,0 +1,2 @@
+export const SN_TAX = 'tax'
+export const TAX_UPDATE_RULES = SN_TAX + '/UPDATE_RULES'
